// models/InteriorSubCategory.ts
type InteriorSubCategory = {
  interiorsubcategoryId: number
  name: string
  catLandingDescription: string
  headline: string
  catLandingDescriptionMarkup: string
  descriptionMarkup: string
  externalUrlActive: boolean
  prettyUrl: string
  externalUrl: string
  fileOrig: string
  static: boolean
  onDisplay: boolean
  subcategoryId: number
}

export default InteriorSubCategory