// models/SubCategory.ts
type SubCategory = {
  subcategoryId: number
  name: string
  catLandingDescription: string
  headling: string
  catLandingDescriptionMarkup: string
  descriptionMarkup: string
  externalUrlActive: string
  prettyUrl: string
  externalUrl: string
  fileOrig: string
  static: boolean
  onDisplay: boolean
  categoryId: number
}

export default SubCategory;